% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_active_awards.R
\name{clean_active_awards}
\alias{clean_active_awards}
\title{Create a clean dataset from Active Awards google sheet (template from USAID Mozambique)}
\usage{
clean_active_awards(ACTIVE_AWARDS_PATH)
}
\arguments{
\item{ACTIVE_AWARDS_PATH}{Path to the Active Awards excel sheet}
}
\value{
A clean dataset with the following columns:sub_sector, activity_name, award_number, total_estimated_cost,
start_date, end_date, u_s_org_local, aor_cor_or_activity_manager,
period, funding_type, pepfar_funding
}
\description{
Create a clean dataset from Active Awards google sheet (template from USAID Mozambique)
}
\examples{
\dontrun{
 df <- process_active_awards("path/to/Active Awards.xlsx")
 }
}
