% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_close_out_tracker.R
\name{clean_close_out_tracker}
\alias{clean_close_out_tracker}
\title{Clean data in the closeout tracker}
\usage{
clean_close_out_tracker(file)
}
\arguments{
\item{file}{closeout tracker file}
}
\value{
a cleaned dataset
}
\description{
Clean data in the closeout tracker
}
\examples{
\dontrun{
 df <- clean_close_out_tracker(file)
 }
}
