% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phoenix_bi_oblg_acc_lines.R
\name{clean_phoenix_bi_oblg_acc_lines}
\alias{clean_phoenix_bi_oblg_acc_lines}
\title{Title: Process Bilateral Obligation Account Lines}
\usage{
clean_phoenix_bi_oblg_acc_lines(file)
}
\arguments{
\item{file}{file containing raw data}
}
\value{
A tibble with the cleaned data
}
\description{
Title: Process Bilateral Obligation Account Lines
}
\examples{
\dontrun{
 df <- clean_phoenix_bi_oblg_acc_lines(file)
 }
}
