% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phoenix_oblg_acc_lines.R
\name{clean_phoenix_oblg_acc_lines}
\alias{clean_phoenix_oblg_acc_lines}
\title{Create a clean dataset from phoneix obligation accounting lines dataset}
\usage{
clean_phoenix_oblg_acc_lines(
  file,
  award_numbers,
  OBLIGATION_TYPE_FILTER,
  DISTRIBUTION_FILTER
)
}
\arguments{
\item{file}{a dataset to clean}

\item{award_numbers}{a vector of award numbers to filter the dataset}

\item{OBLIGATION_TYPE_FILTER}{a vector of obligation types to filter the dataset}

\item{DISTRIBUTION_FILTER}{a vector of distribution to filter the dataset}
}
\value{
a clean dataset
}
\description{
Create a clean dataset from phoneix obligation accounting lines dataset
}
\examples{
\dontrun{
 df <- process_phoenix_oblg_acc_lines(file, award_numbers,
 obligation_type_filter, distribution_filter)
 }
}
