% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phoenix_open_commitments.R
\name{clean_phoenix_open_commitments}
\alias{clean_phoenix_open_commitments}
\title{Clean Phoenix Open Commitments dataset}
\usage{
clean_phoenix_open_commitments(file)
}
\arguments{
\item{file}{A dataset with Phoenix Open Commitments data}
}
\value{
A cleaned dataset
}
\description{
Clean Phoenix Open Commitments dataset
}
\examples{
\dontrun{
 df <- clean_phoenix_open_commitments(file)
 }
}
