% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phoenix_pipeline.R
\name{clean_phoenix_pipeline}
\alias{clean_phoenix_pipeline}
\title{Create cleaned phoenix pipeline dataset with only relevant award numbers}
\usage{
clean_phoenix_pipeline(
  PHOENIX_PIPELINE_PATH,
  all_award_number,
  obligation_type_filter,
  distribution_filter
)
}
\arguments{
\item{PHOENIX_PIPELINE_PATH}{path to phoenix pipeline files}

\item{all_award_number}{list of award numbers}

\item{obligation_type_filter}{list of obligation types}

\item{distribution_filter}{list of distribution codes}
}
\value{
cleaned phoenix pipeline dataset
}
\description{
Create cleaned phoenix pipeline dataset with only relevant award numbers
}
\examples{
\dontrun{
 df <- process_phoenix_pipeline(phoenix_pipeline_path, award_number,
 obligation_type_filter, distribution_filter)
 }
}
