% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_subobligation_summary.R
\name{clean_subobligation_summary}
\alias{clean_subobligation_summary}
\title{Title: Process Subobligation Summary}
\usage{
clean_subobligation_summary(SUBOBLIGATION_SUMMARY_PATH)
}
\arguments{
\item{SUBOBLIGATION_SUMMARY_PATH}{Path to the file containing the raw data}
}
\value{
A tibble with the cleaned data
}
\description{
Title: Process Subobligation Summary
}
\examples{
\dontrun{
 df <- clean_subobligation_summary("path/to/subobligation_summary.xlsx")
 }
}
