% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bi_oblg_acc_lines_updated_all.R
\name{create_bi_oblg_acc_lines_updated_all}
\alias{create_bi_oblg_acc_lines_updated_all}
\title{Function cleans the bi_oblg_acc_lines file that has been updated by the team}
\usage{
create_bi_oblg_acc_lines_updated_all(file, is_pepfar)
}
\arguments{
\item{file}{bi_oblg_acc_lines file that has been updated by team}

\item{is_pepfar}{TRUE or FALSE depending on it if's PEPFAR}
}
\value{
A tibble with the updated bi_oblg_acc_lines
}
\description{
Function cleans the bi_oblg_acc_lines file that has been updated by the team
}
\examples{
\dontrun{
 df <- create_bi_oblg_acc_lines_updated_all(file, TRUE)
 }
}
