% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bi_oblg_acc_lines_updated_mech.R
\name{create_bi_oblg_acc_lines_updated_mech}
\alias{create_bi_oblg_acc_lines_updated_mech}
\title{Creates a dataset from the updated bi_oblg_acc_lines file that has has mechanisms and comments as columns}
\usage{
create_bi_oblg_acc_lines_updated_mech(file, is_pepfar)
}
\arguments{
\item{file}{bi_oblg_acc_lines file that has been updated by team}

\item{is_pepfar}{TRUE or FALSE depending on it if's PEPFAR}
}
\value{
A tibble with the updated bi_oblg_acc_lines
}
\description{
Creates a dataset from the updated bi_oblg_acc_lines file that has has mechanisms and comments as columns
}
\examples{
\dontrun{
 df <- create_bi_oblg_acc_lines_updated_mech(file, TRUE)
 }
}
