% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_history_open_commitments.R
\name{create_history_open_commitments}
\alias{create_history_open_commitments}
\title{Combine all open commitments datasets}
\usage{
create_history_open_commitments(file)
}
\arguments{
\item{file}{A dataset with Phoenix Open Commitments data}
}
\value{
A cleaned dataset
}
\description{
Combine all open commitments datasets
}
\examples{
\dontrun{
 df <- create_history_open_commitments(file)
 }
}
