% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phoenix_transaction_cumulative.R
\name{create_phoenix_transaction_cumulative}
\alias{create_phoenix_transaction_cumulative}
\title{Create cumulative totals to the complete transaction dataset}
\usage{
create_phoenix_transaction_cumulative(phoenix_transaction_df)
}
\arguments{
\item{phoenix_transaction_df}{cleaned phoenix transaction dataset}
}
\value{
a complete transaction dataset with cumulative totals
}
\description{
Create cumulative totals to the complete transaction dataset
}
\examples{
\dontrun{
 df <- create_phoenix_transaction_cumulative(transaction_df)
 }
}
