% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_program_area_name_map}
\alias{data_program_area_name_map}
\title{Variable map for mapping program area to program area name}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_program_area_name_map}{program_area}
\item{data_program_area_name_map}{program_area_name}
}
}
\usage{
data(data_program_area_name_map)
}
\description{
List of indicators
}
\keyword{datasets}
