% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_program_element_map}
\alias{data_program_element_map}
\title{Variable map for mapping old program elements to new program elements}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_program_element_map}{old_program_element}
\item{data_program_element_map}{new_program_area}
}
}
\usage{
data(data_program_element_map)
}
\description{
List of indicators
}
\keyword{datasets}
