% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bilateral_oblg_acc_lines.R
\name{process_bilateral_oblg_acc_lines}
\alias{process_bilateral_oblg_acc_lines}
\title{Title: Process Bilateral Obligation Account Lines}
\usage{
process_bilateral_oblg_acc_lines(path)
}
\arguments{
\item{path}{Path to the file containing the raw data}
}
\value{
A tibble with the cleaned data
}
\description{
Title: Process Bilateral Obligation Account Lines
}
\examples{
\dontrun{
 df <- process_bilateral_oblg_acc_lines(path)
 }
}
