% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phoenix_transaction.R
\name{clean_phoenix_transaction}
\alias{clean_phoenix_transaction}
\title{Process phoenix transaction dataset}
\usage{
clean_phoenix_transaction(file, all_award_number, distribution_filter)
}
\arguments{
\item{file}{path to the file containing the raw data}

\item{all_award_number}{all relevant award numbers}

\item{distribution_filter}{distriction to be included}
}
\value{
cleaned phoenix transaction dataset
}
\description{
Process phoenix transaction dataset
}
\examples{
\dontrun{
 df <- clean_phoenix_transaction(phoenix_transaction_path, award_number,
  distribution_filter)
 }
}
