% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phoenix_transaction.R
\name{clean_phoenix_transaction}
\alias{clean_phoenix_transaction}
\title{Process phoenix transaction dataset}
\usage{
clean_phoenix_transaction(
  PHOENIX_TRANSACTION_PATH,
  all_award_number,
  distribution_filter
)
}
\arguments{
\item{PHOENIX_TRANSACTION_PATH}{folder path to phoneix transaction datasets}

\item{all_award_number}{all award numbers of interest}

\item{distribution_filter}{distriction to be included}
}
\value{
cleaned phoenix transaction dataset
}
\description{
Process phoenix transaction dataset
}
\examples{
\dontrun{
 df <- clean_phoenix_transaction(phoenix_trnsaction_path, award_number,
  distribution_filter)
 }
}
