% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phoenix_transaction.R
\name{clean_phoenix_transaction}
\alias{clean_phoenix_transaction}
\title{Process phoenix transaction dataset}
\usage{
clean_phoenix_transaction(
  file,
  active_award_number,
  distribution_filter,
  frequency = "quarter"
)
}
\arguments{
\item{file}{path to the file containing the raw data}

\item{active_award_number}{active award numbers}

\item{distribution_filter}{distriction to be included}

\item{frequency}{frequency of the data either quarter or month}
}
\value{
cleaned phoenix transaction dataset
}
\description{
Process phoenix transaction dataset
}
\examples{
\dontrun{
 df <- clean_phoenix_transaction(phoenix_trnsaction_path, award_number,
  distribution_filter, frequency = "quarter")
 }
}
