% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_history_bi_oblg_acc_lines.R
\name{create_history_bi_oblg_acc_lines}
\alias{create_history_bi_oblg_acc_lines}
\title{Create history of bi_oblg_acc_lines}
\usage{
create_history_bi_oblg_acc_lines(file, is_pepfar)
}
\arguments{
\item{file}{A dataset with Phoenix Open Commitments data}

\item{is_pepfar}{TRUE if a PEPFAR dataset, FALSE if a non_pepfar dataset}
}
\value{
A cleaned dataset
}
\description{
Create history of bi_oblg_acc_lines
}
\examples{
\dontrun{
 df <- create_history_bi_oblg_acc_lines(file)
 }
}
